Update snapshots and add in Onboard.test.js
