+++ client/src/__tests__/CredentialsForm.test.js
@@ -1,8 +1,6 @@
 /* eslint-disable */
 import React from 'react';
 import '@testing-library/jest-dom/extend-expect';
-import { Grommet } from 'grommet';
-import { theme } from '../styles/themes';
 
 import {
   renderWithRouterAndProviders,
@@ -14,9 +12,14 @@ import CredentialsForm from '../components/Dashboard/CredentialsForm';
 describe('<CredentialsForm />', () => {
   it('should display the component without crashing', () => {
     const helpers = renderWithRouterAndProviders(
-      <Grommet theme={theme}>
-        <CredentialsForm history={{ location: { pathname: '/' } }} />
-      </Grommet>
+      <CredentialsForm history={{ location: { pathname: '/' } }} />
     );
+  });
+
+  it('should match the previous snapshot of the component', () => {
+    const tree = rendererWithRouterAndProviders(
+      <CredentialsForm history={{ location: { pathname: '/' } }} />
+    );
+    expect(tree.toJSON()).toMatchSnapshot();
   });
 });
